fynesse/requirements.txt
ipython-sql
pymysql
geopandas
mlai
yaml
osmnx  # requires matplotlib >=3.5
matplotlib==3.1.3  # must be 3.1.3, otherwise plotting won't work
